feat(user-ui): improve auth modules
