+++ web/gds-user-ui/src/layouts/AuthLayout.tsx
@@ -8,7 +8,7 @@ type AuthLayoutProps = {
 function AuthLayout({ children, ...rest }: AuthLayoutProps) {
   return (
     <Flex fontSize={'xl'} bg={useColorModeValue('white', 'gray.800')} {...rest}>
-      <Stack spacing={8} mx="auto" w="100%" maxW="lg" py={12}>
+      <Stack spacing={8} mx={'auto'} w={'100%'} maxW={'lg'} py={12} px={10}>
         {children}
       </Stack>
     </Flex>
